+++ lib/jbotsim-ui-swing/src/main/java/io/jbotsim/ui/painting/JBackgroundPainter.java
@@ -26,17 +26,62 @@ import io.jbotsim.core.Topology;
 import java.awt.*;
 
 
+/**
+ * <p>The {@link JBackgroundPainter} is JBotSim's default {@link BackgroundPainter} for AWT.</p>
+ *
+ */
 public class JBackgroundPainter implements BackgroundPainter {
     @Override
     public void paintBackground(UIComponent uiComponent, Topology tp) {
         Graphics2D g2d = (Graphics2D) uiComponent.getComponent();
-        g2d.setStroke(new BasicStroke(1));
+        setStroke(g2d, tp);
+        setRenderingHints(g2d, tp);
+        setColor(g2d, tp);
+
         for (Node n : tp.getNodes()) {
-            double sR = n.getSensingRange();
-            if (sR > 0) {
-                g2d.setColor(Color.gray);
-                g2d.drawOval((int) n.getX() - (int) sR, (int) n.getY() - (int) sR, 2 * (int) sR, 2 * (int) sR);
-            }
+            drawSensingRange(g2d, n);
         }
+    }
+
+    protected void drawSensingRange(Graphics2D g2d, Node n) {
+        double sR = n.getSensingRange();
+        if (sR > 0) {
+            g2d.setColor(Color.gray);
+            g2d.drawOval((int) n.getX() - (int) sR, (int) n.getY() - (int) sR, 2 * (int) sR, 2 * (int) sR);
+        }
+    }
+
+
+    /**
+     * <p>Sets the proper {@link java.awt.Color} on the provided {@link Graphics2D} object, with respect to the provided
+     * {@link Topology}.</p>
+     * <p>You can override this method if you need to change the {@link JBackgroundPainter}'s default color management.</p>
+     * @param g2d a {@link Graphics2D} object.
+     * @param topology the associated {@link Topology}.
+     */
+    protected void setColor(Graphics2D g2d, Topology topology) {
+        g2d.setColor(Color.gray);
+    }
+
+    /**
+     * <p>Sets the proper {@link Stroke} on the provided {@link Graphics2D} object, with respect to the provided
+     * {@link Topology}.</p>
+     * <p>You can override this method if you need to change the {@link JBackgroundPainter}'s default {@link Stroke}.</p>
+     * @param g2d a {@link Graphics2D} object.
+     * @param topology the associated {@link Topology}.
+     */
+    protected void setStroke(Graphics2D g2d, Topology topology) {
+        g2d.setStroke(new BasicStroke(1));
+    }
+
+    /**
+     * <p>Sets the proper {@link RenderingHints} on the provided {@link Graphics2D} object, with respect to the provided
+     * {@link Topology}.</p>
+     * <p>You can override this method if you need to change the {@link JBackgroundPainter} rendering behavior.</p>
+     * @param g2d a {@link Graphics2D} object.
+     * @param topology the associated {@link Topology}.
+     */
+    protected void setRenderingHints(Graphics2D g2d, Topology topology) {
+        // default does nothing
     }
 }