+++ lib/jbotsim-ui-swing/src/main/java/io/jbotsim/ui/painting/JNodePainter.java
@@ -25,15 +25,61 @@ import io.jbotsim.ui.JNode;
 
 import java.awt.*;
 
+/**
+ * <p>The {@link JNodePainter} is JBotSim's default {@link NodePainter} for AWT.</p>
+ *
+ */
 public class JNodePainter implements NodePainter {
     @Override
     public void paintNode(UIComponent uiComponent, Node node) {
         Graphics2D g2d = (Graphics2D) uiComponent.getComponent();
+
+        if(node.getColor() == null)
+            return;
+
+        setStroke(g2d, node);
+        setRenderingHints(g2d, node);
+        setColor(g2d, node);
+
+        drawNode(g2d, node);
+    }
+
+    protected void drawNode(Graphics2D g2d, Node node) {
         JNode jn = (JNode) node.getProperty("jnode");
         int drawSize = jn.getWidth() / 2;
-        if (node.getColor() != null) {
-            g2d.setColor(new Color(node.getColor().getRGB()));
-            g2d.fillOval(drawSize / 2, drawSize / 2, drawSize, drawSize);
-        }
+        g2d.fillOval(drawSize / 2, drawSize / 2, drawSize, drawSize);
+    }
+
+    /**
+     * <p>Sets the proper {@link java.awt.Color} on the provided {@link Graphics2D} object, with respect to the provided
+     * {@link Node}.</p>
+     * <p>You can override this method if you need to change the {@link JNodePainter}'s default color management.</p>
+     * @param g2d a {@link Graphics2D} object.
+     * @param node the associated {@link Node}.
+     */
+    protected void setColor(Graphics2D g2d, Node node) {
+        g2d.setColor(new Color(node.getColor().getRGB()));
+    }
+
+    /**
+     * <p>Sets the proper {@link Stroke} on the provided {@link Graphics2D} object, with respect to the provided
+     * {@link Node}.</p>
+     * <p>You can override this method if you need to change the {@link JNodePainter}'s default {@link Stroke}.</p>
+     * @param g2d a {@link Graphics2D} object.
+     * @param node the associated {@link Node}.
+     */
+    protected void setStroke(Graphics2D g2d, Node node) {
+        // default does nothing
+    }
+
+    /**
+     * <p>Sets the proper {@link RenderingHints} on the provided {@link Graphics2D} object, with respect to the provided
+     * {@link Node}.</p>
+     * <p>You can override this method if you need to change the {@link JNodePainter} rendering behavior.</p>
+     * @param g2d a {@link Graphics2D} object.
+     * @param node the associated {@link Node}.
+     */
+    protected void setRenderingHints(Graphics2D g2d, Node node) {
+        // default does nothing
     }
 }